Make two separate hooks for bitcoin and ethereum
Makes two separate requestAccount hook for bitcoin and for ethereum networks.
