+++ src/hooks/ledger-live-app/index.ts
@@ -1,1 +1,3 @@
+export * from "./useRequestAccount"
+export * from "./useRequestBitcoinAccount"
 export * from "./useRequestEthereumAccount"
