+++ src/hooks/ledger-live-app/useRequestBitcoinAccount.ts
@@ -1,0 +1,6 @@
+import { useRequestAccount, UseRequestAccountReturn } from "./useRequestAccount"
+
+export function useRequestBitcoinAccount(): UseRequestAccountReturn {
+  const result = useRequestAccount("Bitcoin")
+  return result
+}
